新增调邮箱，修改UIDatePicker显示，部分代码调整

+++ FaceTalk/facetalk/goodagents/controllers/FTDGoodAgentDetailController.m
@@ -46,7 +46,7 @@
 }
 -(void)menuClickAction:(NSInteger)index
 {
-    NSLog(@"index : %d",index);
+    NSLog(@"index : %ld",(long)index);
 }
