+++ FaceTalk/facetalk/realhelpresult/controller/TFWRealHelpResultController.m
@@ -80,7 +80,7 @@
 
 -(void)menuClickAction:(NSInteger)index
 {
-    NSLog(@"index : %d",index);
+    NSLog(@"index : %ld",(long)index);
 }
 
 - (void)didReceiveMemoryWarning {
